Implement launcher graphic, losing balls, and new ball dev tool
The pinball launcher now has a basic graphic indicating how charged it is.

Added losing balls: When a ball falls off the board, it will be destroyed and a new ball will be brought in from the queue.

The B key will now add a new ball to the queue, for testing purposes.